+++ src/main/java/com/teamharmony/newscommunity/users/controller/UserController.java
@@ -1,10 +1,5 @@
 package com.teamharmony.newscommunity.users.controller;
 
-import com.auth0.jwt.JWT;
-import com.auth0.jwt.JWTVerifier;
-import com.auth0.jwt.algorithms.Algorithm;
-import com.auth0.jwt.interfaces.DecodedJWT;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.teamharmony.newscommunity.users.dto.UserResponseDto;
 import com.teamharmony.newscommunity.users.service.UserService;
 import com.teamharmony.newscommunity.users.vo.ProfileVO;
@@ -14,29 +9,17 @@ import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import org.hibernate.HibernateException;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import javax.validation.constraints.NotBlank;
-import java.io.IOException;
 import java.net.URI;
-import java.util.Date;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
-
-import static org.springframework.http.HttpHeaders.SET_COOKIE;
-import static org.springframework.http.HttpStatus.FORBIDDEN;
-import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
 /**
  * 사용자 관련 요청을 처리
@@ -175,117 +158,6 @@ public class UserController {
 		String username = user.getUsername();
 		URI uri = URI.create(ServletUriComponentsBuilder.fromCurrentContextPath().path("/api/user/update_profile").toUriString());
 		return ResponseEntity.created(uri).body(userService.updateProfile(username, profile));
-	}
-	
-	/**
-	 * 접근 토큰, 갱신 토큰 갱신
-	 *
-	 * @see				UserService#getTokens
-	 * @see				UserService#getUser
-	 * @see				UserService#getRoles
-	 * @see				UserService#updateTokens
-	 */
-	@GetMapping("/token/refresh")
-	public void refreshToken(HttpServletRequest request, HttpServletResponse response) throws IOException {
-		// 클라이언트가 쿠키에 리프레쉬 토큰을 갖고 있는지 확인
-		String refCookie = getRefCookie(request);
-		
-		// 쿠키가 있으면 DB에 있는지 재확인하고 새로운 쿠키를 DB 저장 후 발급
-		if (refCookie != null) {
-			try {
-				Algorithm algorithm = Algorithm.HMAC256("secret".getBytes());
-				JWTVerifier verifier = JWT.require(algorithm)
-				                          .build();
-				DecodedJWT decodedJWT = verifier.verify(refCookie);
-				String username = decodedJWT.getSubject();
-				String allowedToken = userService.getTokens(username).getRefreshToken();
-				if(!allowedToken.equals(refCookie))throw new IllegalArgumentException("Token not found");
-				// Once we get the username, we need to load that user
-				// use the user(loaded using the username) to create a new token( using the refresh token)
-				User user = userService.getUser(username);
-				String access_token = JWT.create()
-				                         .withSubject(user.getUsername())
-				                         .withExpiresAt(new Date(System.currentTimeMillis() + 16*60*60*1000)) // 테스트를 위해 16시간으로 설정
-				                         .withClaim("roles", userService.getRoles(user).stream().map(Role::getName).map(Enum::toString).collect(Collectors.toList()))
-				                         .sign(algorithm);
-				String refresh_token = JWT.create()
-				                          .withSubject(user.getUsername())
-				                          .withExpiresAt(new Date(System.currentTimeMillis() + 7*24*60*60*1000))
-				                          .sign(algorithm);
-				userService.updateTokens(username, access_token, refresh_token);
-				
-				ResponseCookie refresh = ResponseCookie.from("ref_uid", refresh_token)
-				                                       .maxAge(7*24*60*60*1000) // 밀리세컨인지 확인해야됨
-				                                       .httpOnly(true)
-				                                       .secure(true)
-				                                       .sameSite("None")
-				                                       .path("/")
-				                                       .build();
-				response.setHeader(SET_COOKIE, refresh.toString());
-				response.setHeader("token", access_token);
-				response.setContentType(APPLICATION_JSON_VALUE);
-				new ObjectMapper().writeValue(response.getOutputStream(), "success");
-			} catch (Exception e) {
-				removeRefCookie(response);
-				setError(response, e.getMessage());
-			}
-		} else {
-			removeRefCookie(response);
-			setError(response, "Refresh token is missing");
-		}
-	}
-	
-	/**
-	 * 로그아웃
-	 *
-	 * @param 		user 인증된 사용자 정보
-	 * @see				UserService#updateTokens
-	 */
-	@GetMapping("/user/signout")
-	public ResponseEntity<Void>signOut(HttpServletRequest request, HttpServletResponse response, @AuthenticationPrincipal UserDetails user) {
-// 클라이언트가 쿠키에 리프레쉬 토큰을 갖고 있는지 확인
-		String refCookie = getRefCookie(request);
-		// 쿠키가 있으면 삭제
-		if (refCookie != null) {
-			removeRefCookie(response);
-		}
-		// DB에 저장된 토큰 값 공백 처리
-		userService.updateTokens(user.getUsername(), "", "");
-		return ResponseEntity.ok().build();
-	}
-	
-	// 리프레쉬 쿠키값 가져오기
-	private String getRefCookie(HttpServletRequest request) {
-		Cookie[] cookies = request.getCookies();
-		Cookie cookie = null;
-		if (cookies != null && cookies.length > 0) {
-			for (Cookie value : cookies) {
-				if (value.getName()
-				         .equals("ref_uid")) {
-					cookie = value;
-				}
-			}
-		}
-		return cookie.getValue();
-	}
-	
-	// 리프레쉬 쿠키 삭제(만료시간 0으로 설정)
-	private void removeRefCookie(HttpServletResponse response) {
-		ResponseCookie refresh = ResponseCookie.from("ref_uid", "")
-		                                       .maxAge(0)
-		                                       .path("/")
-		                                       .build();
-		response.setHeader(SET_COOKIE, refresh.toString());
-	}
-	
-	// 에러 발생시 헤더와 output 설정
-	private void setError(HttpServletResponse response, String msg) throws IOException {
-		response.setHeader("error", msg);
-		response.setStatus(FORBIDDEN.value());
-		Map<String, String> error = new HashMap<>();
-		error.put("error", msg);
-		response.setContentType(APPLICATION_JSON_VALUE);
-		new ObjectMapper().writeValue(response.getOutputStream(), error);
 	}
 }
