+++ src/main/java/com/teamharmony/newscommunity/users/service/UserService.java
@@ -1,7 +1,5 @@
 package com.teamharmony.newscommunity.users.service;
 
-import com.teamharmony.newscommunity.auth.entity.Tokens;
-import com.teamharmony.newscommunity.auth.repository.TokensRepository;
 import com.teamharmony.newscommunity.users.dto.*;
 import com.teamharmony.newscommunity.users.vo.ProfileVO;
 import com.teamharmony.newscommunity.users.entity.*;
@@ -37,8 +35,6 @@ public class UserService implements UserDetailsService {
 	private final RoleRepository roleRepository;
 	private final UserProfileRepository profileRepository;
 	private final UserRoleRepository userRoleRepository;
-	private final TokensRepository tokensRepository;
-	
 	private final PasswordEncoder passwordEncoder;
 	private final FileStore fileStore;
 	
@@ -101,20 +97,6 @@ public class UserService implements UserDetailsService {
 	public void saveRole(Role role) {
 		log.info("Saving new role {} to the database", role.getName());
 		roleRepository.save(role);
-	}
-	
-	/**
-	 * 토큰 값 변경
-	 *
-	 * @param 		username 해당 사용자 ID
-	 * @param 		access_token 허용된 접근 토큰 값
-	 * @param 		refresh_token 허용된 갱신 토큰 값
-	 * @return 		저장된 토큰 정보
-	 */
-	public Tokens updateTokens(String username, String access_token, String refresh_token) {
-		Tokens tokens = getTokens(username);
-		tokens.update(access_token, refresh_token);
-		return tokensRepository.save(tokens);
 	}
 	
 	/**
@@ -243,20 +225,6 @@ public class UserService implements UserDetailsService {
 	}
 	
 	/**
-	 * 사용자가 지닌 권한 정보 조회
-	 *
-	 * @param 		user 조회할 사용자 ID
-	 * @return 		사용자가 지닌 권한 정보
-	 * @see				UserService#getRoles
-	 */
-	public Collection<Role> getRoles(User user) {
-		Collection<UserRole> userRole = userRoleRepository.findByUser(user);
-		Collection<Role> roles = new ArrayList<>();
-		userRole.forEach(r -> roles.add(r.getRole()));
-		return roles;
-	}
-	
-	/**
 	 * 전체 사용자 조회
 	 *
 	 * @return 		전체 사용자 정보
@@ -284,17 +252,6 @@ public class UserService implements UserDetailsService {
 		body.put("link", getProfileImageUrl(username));
 		body.put("profile", profileDto);
 		return body;
-	}
-	
-	/**
-	 * 사용자의 허용 토큰 정보 조회
-	 *
-	 * @param 		username 조회할 사용자 ID
-	 * @return 		사용자의 허용 토큰 정보
-	 */
-	public Tokens getTokens(String username) {
-		log.info("Fetching tokens of user {}", username);
-		return tokensRepository.findByUsername(username);
 	}
 	
 	/**
